rearrange: create navigation module
