+++ src/helpers/listener.js
@@ -1,41 +1,11 @@
-import {resolve, join} from 'path';
-import { chdir } from 'process';
+import { resolve, join } from 'path';
 import { getCurrentDirectory } from './console.js';
 import { open, readdir, rename, stat, unlink } from 'fs/promises';
 import { createReadStream, createWriteStream } from 'fs';
 import { pipeline } from 'stream/promises';
 import { output } from './output.js';
 
-const up = (path) => {
-  try {
-    if (path) {
-      console.log('Invalid input');
-    } else {
-      const currentDir = process.cwd();
-      const parentDir = resolve(currentDir, '..');
-      if (parentDir !== currentDir) {
-        process.chdir(parentDir);
-      }
-      getCurrentDirectory();
-    }
-  } catch {
-    console.log('Operation failed');
-  }
-};
-
-const cd = (path) => {
-  try {
-    if (path) {
-      const currentDir = process.cwd();
-      chdir(resolve(currentDir, path));
-      getCurrentDirectory();
-    } else {
-      console.log('Invalid input');
-    }
-  } catch {
-    console.log('Operation failed');
-  }
-}
+import { up, cd } from '../operations/navigation.js';
 
 const ls = async (path) => {
   try {
